added support for editing multiple objects at once
Sort of crazy hackery, where we grab all the properties a list of
objects have in common, emit a new class in IL with those properties and
a notification on set.  An instance of this new class is passed to the
propertygrid, on set it modifies all the original objects accordingly.